+++ R/utils-autolayer.R
@@ -1,0 +1,9 @@
+#' Autolayer generic function
+#'
+#' See [ggplot2::autolayer()], [autolayer.barrier()] for details.
+#'
+#' @name autolayer
+#' @keywords internal
+#' @export
+#' @importFrom ggplot2 autolayer
+NULL